requirements.txt
pandas>=1.3
python-dotenv>=0.21
pytest>=7.0
pytest-cov>=4.0